// src/types/types.ts
export interface Media {
  name: string;
  resource_type: string;
  resource_value: string;
  thumbnail_url?: string;
}

export interface ChecklistItem {
  color: string;
  icon: string;
  id: string;
  list_page_visibility: boolean;
  text: string;
}

export interface SEOMeta {
  content: string;
  type: string;
  value: string;
}

export interface SEOSchema {
  meta_name: string;
  meta_value: string;
  type: string;
}

export interface SEO {
  defaultMeta: SEOMeta[];
  description: string;
  keywords: string[];
  schema: SEOSchema[];
  title: string;
}

export interface CTAText {
  name: string;
  value: string;
}

export interface SectionValue {
  background_color?: string;
  background_img?: string;
  checklist_text_color?: string;
  end_at?: string;
  id?: string;
  start_at?: string;
  template?: string;
  text?: string;
  description?: string;
  has_instructor_page?: boolean;
  image?: string;
  name?: string;
  short_description?: string;
  slug?: string;
  icon?: string;
  subtitle?: string;
  title?: string;
  color?: string;
  background?: {
    image: string;
    primary_color: string;
    secondary_color: string;
  };
  cta?: {
    clicked_url: string;
    color: string;
    text: string;
  };
  description_color?: string;
  thumbnail?: string;
  title_color?: string;
  top_left_icon_img?: string;
  checklist?: string[];
  file_type?: string;
  file_url?: string;
  video_thumbnail?: string;
  profile_image?: string;
  testimonial?: string;
  thumb?: string;
  video_type?: string;
  video_url?: string;
  answer?: string;
  question?: string;
}

export interface Section {
  type: string;
  name: string;
  description: string;
  bg_color: string;
  order_idx: number;
  values: SectionValue[];
}

export interface OldInfo {
  cat_id: number;
  course_id: number;
  platform: string;
  skills_cat_id: number;
  slug: string;
}

export interface CourseData {
  slug: string;
  id: number;
  title: string;
  description: string;
  platform: string;
  type: string;
  modality: string;
  old_info: OldInfo;
  start_at: string;
  media: Media[];
  checklist: ChecklistItem[];
  seo: SEO;
  cta_text: CTAText;
  sections: Section[];
  is_cohort_based_course: boolean;
  secondary_cta_group: any[];
  delivery_method: string;
}

export interface IELTSCourseData {
  code: number;
  data: CourseData;
  error: any[];
  message: string;
  payload: any[];
  status_code: number;
}